@@ -12,22 +12,24 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.2",
-    "@agm/snazzy-info-window": "^1.0.0-beta.2",
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@angular/service-worker": "^5.2.3",
-    "@ngx-loading-bar/http-client": "^1.3.0",
-    "@ngx-loading-bar/router": "^1.3.0",
+    "@agm/core": "^1.0.0-beta.3",
+    "@agm/snazzy-info-window": "^1.0.0-beta.3",
+    "@angular/animations": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/router": "6.0.3",
+    "@angular/service-worker": "6.0.3",
+    "@ngx-loading-bar/http-client": "^2.1.0",
+    "@ngx-loading-bar/router": "^2.1.0",
     "@types/googlemaps": "^3.30.8",
-    "angular-bootstrap-md": "^5.0.5",
+    "ag-grid": "^17.1.1",
+    "ag-grid-angular": "^17.1.0",
+    "angular-bootstrap-md": "^6.1.2",
     "bootstrap": "^4.0.0-alpha.6",
     "chart.js": "^2.5.0",
     "core-js": "^2.4.1",
@@ -41,16 +43,16 @@
     "ng2-cookies": "^1.0.12",
     "pouchdb": "^6.4.3",
     "pouchdb-browser": "^6.4.3",
-    "rxjs": "^5.5.2",
+    "rxjs": "^6.2.0",
     "snazzy-info-window": "^1.1.1",
     "tether": "^1.4.3",
     "topojson": "^3.0.2",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.7",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular/cli": "6.0.7",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/language-service": "6.0.3",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/jquery": "^3.2.16",
@@ -68,6 +70,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.6"
   }
 }
