@@ -2,8 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, Injectable } from '@angular/core';
 import { HttpClientModule, HttpEvent, HttpInterceptor, HttpHandler, HttpRequest, HTTP_INTERCEPTORS} from '@angular/common/http';
 import { RouterModule, Routes} from '@angular/router';
-import { LoadingBarHttpClientModule } from '@ngx-loading-bar/http-client';
-import { LoadingBarRouterModule } from '@ngx-loading-bar/router';
+
 import { MDBBootstrapModule } from 'angular-bootstrap-md';
 import { AgmCoreModule } from '@agm/core';
 
@@ -19,6 +18,8 @@ import { routing } from './app.routing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Exception404Component } from './exception404/exception404.component';
+import { LoadingBarHttpClientModule } from '@ngx-loading-bar/http-client';
+import { LoadingBarRouterModule } from '@ngx-loading-bar/router';
 
 @Injectable()
 export class XhrInterceptor implements HttpInterceptor {
@@ -52,9 +53,10 @@ export class XhrInterceptor implements HttpInterceptor {
     AgmCoreModule.forRoot({
       apiKey:'AIzaSyDMIi1qnl_K9VYeOx_c1mHttcTirQV2Jvc'
     }),
-    LoadingBarRouterModule,
-    LoadingBarHttpClientModule,
-    MDBBootstrapModule.forRoot()
+    
+    MDBBootstrapModule.forRoot(),
+   // LoadingBarHttpClientModule,
+    LoadingBarRouterModule
   ],
   schemas: [NO_ERRORS_SCHEMA],
   providers: [],
